Removed a bug concerning NIM_SETVERSION.
We call Shell_NotifyIcon(NIM_SETVERSION, &nid) with
nid.uVersion set to NOTIFYICON_VERSION_4.
The call returns TRUE indicating success, yet it doesn't
change the icon behavior like it should. The icon does
not send WM_CONTEXTMENU on being right-clicked and the
wParam is not being set to the mouse coordinates.

Thus, we stop relying on NIM_SERVERSION. We query the
mouse position ourselves and answer to WM_RBUTTONUP instead
of WM_CONTEXTMENU.