Removed Xlib type Window from public interface of XDo.
X11/X.h pollutes the global namespace and thus must not be included
in a header file.
Because of that, we may not use any Xlib types in the public
interface of class core::XDo.
Thus, the public member functions of core::XDo have been modified
to not use the Window type anymore.